refactor: inline the crypto routing into the JWT orchestrators, drop ICryptoRouter
Each router method had a single caller, so the separate type and its interface
did not earn their keep. Move the signing routing back into JsonWebTokenSigner
and the JWE key-management routing (encrypt, decrypt, and the external ECDH-ES
agreement) back into JsonWebTokenEncryptor, and delete ICryptoRouter and
CryptoRouter. The optional external ports (IExternalSigner, IExternalKeyEncryptor)
become constructor dependencies of the orchestrators that use them.

Behaviour is byte-identical: the in-process and external paths are unchanged,
and the full suite (including the known-answer vectors, the Microsoft.IdentityModel
interop oracle, and the external signer/key-encryptor tests) stays green.
