+++ Abblix.Jwt/JsonWebTokenEncryptor.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics.CodeAnalysis;
+using System.Security.Cryptography;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Nodes;
@@ -14,10 +15,14 @@ namespace Abblix.Jwt;
 /// Handles encryption and decryption of JSON Web Encryption (JWE) tokens.
 /// Implements RFC 7516 (JWE) encryption and decryption.
 /// </summary>
-/// <param name="serviceProvider">The service provider for resolving content encryptors by algorithm.</param>
-/// <param name="router">The crypto router that dispatches key management in process or to an external key
-/// custodian and owns the fail-closed decision.</param>
-internal class JsonWebTokenEncryptor(IServiceProvider serviceProvider, ICryptoRouter router) : IJsonWebTokenEncryptor
+/// <param name="serviceProvider">The service provider for resolving the keyed content encryptors and key
+/// encryptors by algorithm.</param>
+/// <param name="externalKeyEncryptor">Optional host port for JWE key management with an external key
+/// custodian (RSA/symmetric unwrap, symmetric wrap, ECDH agreement) when a key is published public-only.
+/// Absent (null) means no external encryption keys.</param>
+internal class JsonWebTokenEncryptor(
+    IServiceProvider serviceProvider,
+    IExternalKeyEncryptor? externalKeyEncryptor = null) : IJsonWebTokenEncryptor
 {
     /// <summary>
     /// Encrypts an inner JWS token to create a JWE token.
@@ -31,9 +36,9 @@ internal class JsonWebTokenEncryptor(IServiceProvider serviceProvider, ICryptoRo
         string contentEncryptionAlgorithm,
         CancellationToken cancellationToken = default)
     {
-        // Key validation and the local-vs-external fail-closed decision live in the router. The public-key
-        // guard that used to sit here would reject an external symmetric key (which has no public half), so
-        // it moves into the router, where the routing knows what each key can actually do.
+        // Key validation and the local-vs-external fail-closed decision live in EncryptKeyAsync below. The
+        // public-key guard that used to sit here would reject an external symmetric key (which has no public
+        // half), so it lives with the key-management routing, which knows what each key can actually do.
 
         // Resolve content encryptor to get required CEK size
         var contentEncryptor = serviceProvider.GetRequiredKeyedService<IDataEncryptor>(contentEncryptionAlgorithm);
@@ -46,11 +51,11 @@ internal class JsonWebTokenEncryptor(IServiceProvider serviceProvider, ICryptoRo
             KeyId = encryptionKey.KeyId
         };
 
-        // The router produces the CEK and protects it, in process or via an external custodian. Key-wrapping
-        // algorithms return a random CEK; "dir" returns the shared key itself; ECDH-ES derives the CEK from
-        // the ephemeral-static agreement. Either step may add algorithm parameters to the header ('epk' for
+        // Produce the CEK and protect it, in process or via an external custodian. Key-wrapping algorithms
+        // return a random CEK; "dir" returns the shared key itself; ECDH-ES derives the CEK from the
+        // ephemeral-static agreement. Either step may add algorithm parameters to the header ('epk' for
         // ECDH-ES, 'iv'/'tag' for AES-GCM key wrap, 'p2s'/'p2c' for PBES2).
-        var (cek, encryptedKey) = await router.EncryptKeyAsync(
+        var (cek, encryptedKey) = await EncryptKeyAsync(
             header, encryptionKey, keyEncryptionAlgorithm, contentEncryptor.KeySizeInBytes, cancellationToken);
 
         // Encode header AFTER key encryption (in case it was modified)
@@ -163,9 +168,9 @@ internal class JsonWebTokenEncryptor(IServiceProvider serviceProvider, ICryptoRo
         {
             keyFound = true;
 
-            // The router recovers the CEK - in process for a local key, or via the external custodian for a
-            // public-only key - and returns null on any decryption failure, which the mitigation below turns
-            // into a uniform outcome. The "must have private material" gate now lives in the router.
+            // DecryptKeyAsync recovers the CEK - in process for a local key, or via the external custodian
+            // for a public-only key - and returns null on any decryption failure, which the mitigation below
+            // turns into a uniform outcome. The "must have private material" gate lives with that routing.
 
             // RFC 7516 §11.5: substitute a randomly generated CEK of the correct size and still run the
             // AEAD step when CEK decryption fails — wrong key, malformed padding, or an unregistered
@@ -178,7 +183,7 @@ internal class JsonWebTokenEncryptor(IServiceProvider serviceProvider, ICryptoRo
             // This is what makes RSA1_5 (RSAES-PKCS1-v1_5) safe to support: it closes the
             // Bleichenbacher/Manger padding oracle by removing the observable difference between valid
             // and invalid padding.
-            var contentEncryptionKey = await router.DecryptKeyAsync(header, key, algorithm, encryptedKey, cancellationToken);
+            var contentEncryptionKey = await DecryptKeyAsync(header, key, algorithm, encryptedKey, cancellationToken);
             if (contentEncryptionKey == null || contentEncryptionKey.Length != contentDecryptor.KeySizeInBytes)
                 contentEncryptionKey = CryptoRandom.GetRandomBytes(contentDecryptor.KeySizeInBytes);
 
@@ -195,4 +200,242 @@ internal class JsonWebTokenEncryptor(IServiceProvider serviceProvider, ICryptoRo
             keyFound ? "Failed to decrypt JWE with any available key" : "No decryption keys found");
     }
 
+    /// <summary>
+    /// Produces the Content Encryption Key and its wrapped form. Asymmetric key management (RSA, ECDH-ES)
+    /// encrypts with the recipient's public half, so it runs in process even for an external key; a symmetric
+    /// key whose secret is absent is wrapped by the external custodian; direct and password-based key
+    /// management have no external form and fail closed.
+    /// </summary>
+    private async ValueTask<(byte[] contentEncryptionKey, byte[] encryptedKey)> EncryptKeyAsync(
+        JsonWebTokenHeader header,
+        JsonWebKey encryptionKey,
+        string algorithm,
+        int contentKeySizeInBytes,
+        CancellationToken cancellationToken)
+    {
+        // Asymmetric key management (RSA, ECDH-ES) encrypts with the recipient's public half, so it runs in
+        // process even for an external key - the public half is present. Only a symmetric key whose secret
+        // bytes are absent must be wrapped by the external custodian.
+        if (encryptionKey is OctetJsonWebKey { HasPrivateKey: false })
+        {
+            if (!IsExternallyWrappable(algorithm))
+                throw NoExternalForm(encryptionKey, algorithm);
+
+            var port = externalKeyEncryptor ?? throw NoKeyEncryptorPort(encryptionKey);
+            var cek = CryptoRandom.GetRandomBytes(contentKeySizeInBytes);
+            var wrapped = await port.WrapKeyAsync(RequireKid(encryptionKey), algorithm, header, cek, cancellationToken);
+            return (cek, wrapped);
+        }
+
+        // In-process dispatch to the keyed primitive.
+        return EncryptLocally(encryptionKey);
+
+        (byte[], byte[]) EncryptLocally(JsonWebKey key) => key switch
+        {
+            RsaJsonWebKey rsaKey => EncryptBy(rsaKey),
+            OctetJsonWebKey octetKey => EncryptBy(octetKey),
+            EllipticCurveJsonWebKey ecKey => EncryptBy(ecKey),
+            _ => throw new InvalidOperationException($"No key encryptor registered for key type: {key.GetType().Name}"),
+        };
+
+        (byte[], byte[]) EncryptBy<TJsonWebKey>(TJsonWebKey jwk) where TJsonWebKey : JsonWebKey
+        {
+            var keyEncryptor = serviceProvider.GetRequiredKeyedService<IKeyEncryptor<TJsonWebKey>>(algorithm);
+            var cek = keyEncryptor.GenerateContentEncryptionKey(header, jwk, contentKeySizeInBytes);
+            return (cek, keyEncryptor.EncryptKey(header, jwk, cek));
+        }
+    }
+
+    /// <summary>
+    /// Recovers the Content Encryption Key from a JWE Encrypted Key. A key with secret material unwraps in
+    /// process; a public-only key routes the private operation (RSA decrypt, ECDH agreement, symmetric
+    /// unwrap) to an external custodian. Returns null on any decryption failure so a wrong key is
+    /// indistinguishable from a bad ciphertext (the RFC 7516 §11.5 mitigation upstream relies on this) and
+    /// never throws on the attacker-supplied header.
+    /// </summary>
+    private async ValueTask<byte[]?> DecryptKeyAsync(
+        JsonWebTokenHeader header,
+        JsonWebKey decryptionKey,
+        string algorithm,
+        byte[] encryptedKey,
+        CancellationToken cancellationToken)
+    {
+        // Secret material present: unwrap in process with the keyed primitive (the unchanged path). This
+        // branch is the "must have private material" gate.
+        if (decryptionKey.HasPrivateKey)
+            return DecryptLocally(decryptionKey);
+
+        // Public-only key: route the private operation to the external custodian by kid. Here 'algorithm'
+        // and 'kid' come from the attacker-supplied JWE header, so anything this cannot handle returns null
+        // and the caller substitutes a random CEK for a uniform failure (RFC 7516 §11.5). A throw would be an
+        // oracle or a DoS; a genuinely missing port is a misconfiguration the startup guard rejects at boot,
+        // not at request time.
+        if (externalKeyEncryptor is not { } port || decryptionKey.KeyId is not { } kid)
+            return null;
+
+        return decryptionKey switch
+        {
+            // RSA decrypt and symmetric unwrap are single remote calls. The algorithm must match the key
+            // type, mirroring the keyed-DI validation the in-process path gets for free.
+            RsaJsonWebKey when IsExternalRsaAlgorithm(algorithm)
+                => await port.UnwrapKeyAsync(kid, algorithm, header, encryptedKey, cancellationToken),
+            OctetJsonWebKey when IsExternallyWrappable(algorithm)
+                => await port.UnwrapKeyAsync(kid, algorithm, header, encryptedKey, cancellationToken),
+
+            // ECDH-ES: only the agreement needs the private key, so only it is remote; the KDF and any AES
+            // key unwrap run locally on the returned shared secret.
+            EllipticCurveJsonWebKey ecKey when IsEcdhEsAlgorithm(algorithm) => await AgreeExternallyAsync(
+                header, ecKey, algorithm, encryptedKey, port, kid, cancellationToken),
+
+            // dir / PBES2, or an algorithm that does not match the key type: no external form, uniform null.
+            _ => null,
+        };
+
+        byte[]? DecryptLocally(JsonWebKey key) => key switch
+        {
+            RsaJsonWebKey rsaKey => TryDecryptBy(rsaKey),
+            OctetJsonWebKey octetKey => TryDecryptBy(octetKey),
+            EllipticCurveJsonWebKey ecKey => TryDecryptBy(ecKey),
+            _ => null,
+        };
+
+        byte[]? TryDecryptBy<TJsonWebKey>(TJsonWebKey jwk) where TJsonWebKey : JsonWebKey
+        {
+            var keyEncryptor = serviceProvider.GetKeyedService<IKeyEncryptor<TJsonWebKey>>(algorithm);
+            return keyEncryptor != null && keyEncryptor.TryDecryptKey(header, jwk, encryptedKey, out var cek)
+                ? cek
+                : null;
+        }
+    }
+
+    /// <summary>
+    /// External ECDH-ES: the custodian performs the agreement with the recipient's private key and returns
+    /// the raw shared secret Z; the Concat KDF and any AES key unwrap run locally. Mirrors the guards of the
+    /// in-process EcdhEsKeyEncryptor.TryDecryptKey - only the agreement step is remote.
+    /// </summary>
+    private async ValueTask<byte[]?> AgreeExternallyAsync(
+        JsonWebTokenHeader header,
+        EllipticCurveJsonWebKey recipientKey,
+        string algorithm,
+        byte[] encryptedKey,
+        IExternalKeyEncryptor port,
+        string kid,
+        CancellationToken cancellationToken)
+    {
+        // 'epk', 'apu' and 'apv' come from the attacker-supplied header, so a malformed value fails as a
+        // uniform null (mirroring the in-process EcdhEsKeyEncryptor.TryDecryptKey) rather than throwing.
+        try
+        {
+            // RFC 7518 §4.6.2: when both 'apu' and 'apv' are present they must differ, otherwise the
+            // producer and recipient identities collapse and the KDF binding loses meaning.
+            var apu = header.AgreementPartyUInfo;
+            var apv = header.AgreementPartyVInfo;
+            if (apu != null && apv != null && string.Equals(apu, apv, StringComparison.Ordinal))
+                return null;
+
+            // The originator's ephemeral public key is mandatory and must live on the recipient's curve.
+            if (header.EphemeralPublicKey is not EllipticCurveJsonWebKey { HasPublicKey: true } ephemeralKey
+                || !string.Equals(ephemeralKey.Curve, recipientKey.Curve, StringComparison.Ordinal))
+                return null;
+
+            // Z is the raw ECDH shared secret (NIST SP 800-56A / RFC 7518 §4.6); here it comes from the
+            // custodian rather than an in-process agreement, but the KDF over it is identical.
+            var sharedSecretZ = await port.AgreeKeyAsync(kid, algorithm, ephemeralKey, cancellationToken);
+            try
+            {
+                if (KeyWrapSize(algorithm) is { } kekSize)
+                {
+                    var kek = ConcatKeyDerivation.DeriveKey(sharedSecretZ, algorithm, apu, apv, kekSize);
+                    return AesKeyWrap.TryUnwrap(kek, encryptedKey, out var cek) ? cek : null;
+                }
+
+                // Direct Key Agreement: the encrypted key must be empty and the derived key IS the CEK, sized
+                // for the content encryption algorithm named by 'enc'.
+                if (encryptedKey.Length != 0)
+                    return null;
+
+                if (header.EncryptionAlgorithm is not { } contentEncryptionAlgorithm
+                    || serviceProvider.GetKeyedService<IDataEncryptor>(contentEncryptionAlgorithm) is not { } contentEncryptor)
+                    return null;
+
+                return ConcatKeyDerivation.DeriveKey(
+                    sharedSecretZ, contentEncryptionAlgorithm, apu, apv, contentEncryptor.KeySizeInBytes);
+            }
+            finally
+            {
+                CryptographicOperations.ZeroMemory(sharedSecretZ);
+            }
+        }
+        catch (JsonException)
+        {
+            return null; // malformed 'epk'
+        }
+        catch (FormatException)
+        {
+            return null; // malformed base64url in 'apu' / 'apv'
+        }
+        catch (CryptographicException)
+        {
+            return null; // agreement or key-unwrap failure
+        }
+    }
+
+    /// <summary>
+    /// The symmetric key-management algorithms whose wrap/unwrap an external custodian can perform. Direct
+    /// encryption (dir) and PBES2 are excluded by construction: dir's CEK is the shared secret itself, and
+    /// PBES2 derives the KEK from the secret by a password KDF, so neither has a remote wrap/unwrap form.
+    /// </summary>
+    private static bool IsExternallyWrappable(string algorithm) => algorithm switch
+    {
+        EncryptionAlgorithms.KeyManagement.Aes128KW or
+        EncryptionAlgorithms.KeyManagement.Aes192KW or
+        EncryptionAlgorithms.KeyManagement.Aes256KW or
+        EncryptionAlgorithms.KeyManagement.Aes128Gcmkw or
+        EncryptionAlgorithms.KeyManagement.Aes192Gcmkw or
+        EncryptionAlgorithms.KeyManagement.Aes256Gcmkw => true,
+        _ => false,
+    };
+
+    /// <summary>The RSA key-transport algorithms an external RSA key can decrypt.</summary>
+    private static bool IsExternalRsaAlgorithm(string algorithm) => algorithm switch
+    {
+        EncryptionAlgorithms.KeyManagement.RsaOaep or
+        EncryptionAlgorithms.KeyManagement.RsaOaep256 or
+        EncryptionAlgorithms.KeyManagement.Rsa1_5 => true,
+        _ => false,
+    };
+
+    /// <summary>The ECDH-ES family an external EC key can agree.</summary>
+    private static bool IsEcdhEsAlgorithm(string algorithm) => algorithm switch
+    {
+        EncryptionAlgorithms.KeyManagement.EcdhEs or
+        EncryptionAlgorithms.KeyManagement.EcdhEsAes128KW or
+        EncryptionAlgorithms.KeyManagement.EcdhEsAes192KW or
+        EncryptionAlgorithms.KeyManagement.EcdhEsAes256KW => true,
+        _ => false,
+    };
+
+    /// <summary>
+    /// The RFC 3394 KEK size for the ECDH-ES key-wrapping variants, or null for Direct Key Agreement where
+    /// the derived key is the CEK itself. Mirrors EcdhEsKeyEncryptor's mode selection for the remote path.
+    /// </summary>
+    private static int? KeyWrapSize(string algorithm) => algorithm switch
+    {
+        EncryptionAlgorithms.KeyManagement.EcdhEsAes128KW => 16,
+        EncryptionAlgorithms.KeyManagement.EcdhEsAes192KW => 24,
+        EncryptionAlgorithms.KeyManagement.EcdhEsAes256KW => 32,
+        _ => null,
+    };
+
+    private static string RequireKid(JsonWebKey key)
+        => key.KeyId ?? throw new InvalidOperationException(
+            "An external key must carry a 'kid': it is the key custodian's handle.");
+
+    private static InvalidOperationException NoKeyEncryptorPort(JsonWebKey key)
+        => new($"Key (kid={key.KeyId}) has no secret material and no {nameof(IExternalKeyEncryptor)} is " +
+               "configured to serve it from an external custodian.");
+
+    private static InvalidOperationException NoExternalForm(JsonWebKey key, string algorithm)
+        => new($"Key (kid={key.KeyId}) is external, but the key-management algorithm '{algorithm}' has no " +
+               "external form (direct and password-based key management cannot be externalised); failing closed.");
 }
