+++ Abblix.Jwt/JsonWebTokenSigner.cs
@@ -13,10 +13,11 @@ namespace Abblix.Jwt;
 /// <summary>
 /// Handles signing and signature verification of JSON Web Signature (JWS) tokens.
 /// </summary>
-/// <param name="serviceProvider">The service provider for resolving signature verifiers by algorithm
-/// (the verify path stays local against the public key; signing is routed via <paramref name="router"/>).</param>
-/// <param name="router">Routes signing to the in-process keyed primitive or, in later work, an external
-/// key custodian, and owns the fail-closed "signing requires the private half" decision.</param>
+/// <param name="serviceProvider">The service provider for resolving the keyed signature primitives
+/// (<see cref="IDataSigner{TJsonWebKey}"/>) by algorithm, for both signing and verification.</param>
+/// <param name="externalSigner">Optional host port that signs with an external key custodian (HSM/KMS/
+/// vault) when a signing key is published public-only. Absent (null) means no external signing keys - the
+/// optional dependency defaults to null, so the container passes null when the host registers no port.</param>
 /// <param name="logger">Records signature-verification outcomes as structured events. The
 /// caller-facing <see cref="JwtValidationError"/> carries a human-readable description, but
 /// FAPI 2.0 audit-logging requires a granular event-type on every key-resolution failure
@@ -25,7 +26,7 @@ namespace Abblix.Jwt;
 internal partial class JsonWebTokenSigner(
     ILogger<JsonWebTokenSigner> logger,
     IServiceProvider serviceProvider,
-    ICryptoRouter router) : IJsonWebTokenSigner
+    IExternalSigner? externalSigner = null) : IJsonWebTokenSigner
 {
     private static readonly JsonSerializerOptions Options = new() { WriteIndented = false };
 
@@ -79,12 +80,51 @@ internal partial class JsonWebTokenSigner(
         var signingInput = $"{EncodeJson(token.Header.Json)}.{EncodeJson(token.Payload.Json)}";
         var signingBytes = Encoding.UTF8.GetBytes(signingInput);
 
-        // The router owns the fail-closed "signing requires the private half" decision (it is what
-        // selects the in-process keyed signer vs. an external key custodian), so the key-material guard
-        // and the key-type dispatch that used to live here now sit in one place behind ICryptoRouter.
-        var signature = await router.SignAsync(signingKey, algorithm, signingBytes, cancellationToken);
+        // Sign with the private half in process when present; otherwise route to the external key custodian
+        // by kid (the absence of private material, not a flag, selects the remote path); fail closed when a
+        // public-only key has no external signer.
+        var signature = await SignBytesAsync(signingKey, algorithm, signingBytes, cancellationToken);
 
         return $"{signingInput}.{Base64Url.EncodeToString(signature)}";
+    }
+
+    /// <summary>
+    /// Produces the signature bytes: in process with the keyed <see cref="IDataSigner{TJsonWebKey}"/> when
+    /// the key carries private material, via the external custodian by kid when it does not, else fail closed.
+    /// </summary>
+    private ValueTask<byte[]> SignBytesAsync(
+        JsonWebKey signingKey, string algorithm, byte[] data, CancellationToken cancellationToken)
+    {
+        if (signingKey.HasPrivateKey)
+            return new ValueTask<byte[]>(SignLocally(signingKey));
+
+        if (externalSigner != null)
+        {
+            // The kid published in the token and JWKS IS the custodian's handle - no separate identifier
+            // and no mapping - so an external key must carry one.
+            var kid = signingKey.KeyId ?? throw new InvalidOperationException(
+                "An external signing key must carry a 'kid': it is the key custodian's handle.");
+
+            return externalSigner.SignAsync(kid, algorithm, data, cancellationToken);
+        }
+
+        throw new InvalidOperationException(
+            $"Signing requires private key material for key (kid={signingKey.KeyId}); it carries none " +
+            "and no external signer is configured.");
+
+        byte[] SignLocally(JsonWebKey key) => key switch
+        {
+            RsaJsonWebKey rsaKey => SignBy(rsaKey),
+            EllipticCurveJsonWebKey ecKey => SignBy(ecKey),
+            OctetJsonWebKey octetKey => SignBy(octetKey),
+            _ => throw new InvalidOperationException($"No signer registered for key type: {key.GetType().Name}"),
+        };
+
+        byte[] SignBy<TJsonWebKey>(TJsonWebKey jwk) where TJsonWebKey : JsonWebKey
+        {
+            var dataSigner = serviceProvider.GetRequiredKeyedService<IDataSigner<TJsonWebKey>>(algorithm);
+            return dataSigner.Sign(jwk, data);
+        }
     }
 
     /// <summary>
